requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiogram.contrib==1.1.4
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
altgraph==0.17.4
annotated-types==0.7.0
asyncio==3.4.3
attrs==25.3.0
autoflake==2.3.1
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
defusedxml==0.7.1
dnspython==2.7.0
emoji==2.14.1
environs==14.1.1
executor==23.2
fake-useragent==2.1.0
Faker==37.1.0
fasteners==0.19
flake8==7.2.0
Flask==3.1.0
frozenlist==1.5.0
google-search==1.1.1
google_search_results==2.4.2
googlesearch-python==1.3.0
greenlet==3.2.1
h2==4.2.0
hpack==4.1.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ipwhois==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
magic-filter==1.0.12
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
multidict==6.3.2
mypy_extensions==1.1.0
opentele==1.15.1
packaging==24.2
pathspec==0.12.1
peewee==3.17.9
pefile==2023.2.7
phonenumbers==9.0.2
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
property-manager==3.0
psutil==7.0.0
py-whois==0.1
pyaes==1.6.1
pyasn1==0.6.1
pycodestyle==2.13.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.3.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
Pyrogram==2.0.106
PySocks==1.7.1
pystyle==2.9
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nmap==0.7.1
python-whois==0.9.5
pythonwhois==2.4.3
pywhois==0.1
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9
setuptools==78.1.0
six==1.17.0
socialscan==2.0.1
soupsieve==2.6
SQLAlchemy==2.0.40
telebot==0.0.5
Telethon==1.39.0
termcolor==3.0.1
TgCrypto==1.2.5
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
verboselogs==1.7
Werkzeug==3.1.3
wheel==0.45.1
whois==1.20240129.2
yarl==1.19.0
